# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory/bootstrap:
    post:
      operationId: bootstrapInventory
      tags: [inventory]
      summary: Seed the inventory database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryBootstrapInput"
      responses:
        "200":
          description: Current inventory state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryState"
  /inventory/items:
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create an inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/import:
    post:
      operationId: importInventoryItems
      tags: [inventory]
      summary: Bulk import inventory items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryImportInput"
      responses:
        "201":
          description: Imported inventory items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryImportResponse"
        "409":
          description: One or more item IDs already exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inventory/items/{id}:
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - $ref: "#/components/parameters/InventoryItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - $ref: "#/components/parameters/InventoryItemId"
      responses:
        "204":
          description: Inventory item deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /inventory/adjustments:
    post:
      operationId: adjustInventoryQuantity
      tags: [inventory]
      summary: Set an inventory quantity and record stock activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryAdjustmentInput"
      responses:
        "200":
          description: Updated stock activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockActivity"
        "404":
          $ref: "#/components/responses/NotFound"
  /inventory/audit-records:
    post:
      operationId: createAuditRecord
      tags: [inventory]
      summary: Record a material request and deduct inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditRecordInput"
      responses:
        "201":
          description: Created audit record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditRecord"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Not enough inventory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inventory/audit-records/{id}:
    delete:
      operationId: deleteAuditRecord
      tags: [inventory]
      summary: Delete an audit record and restore its quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Audit record deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /inventory/borrowed-items:
    post:
      operationId: createBorrowedItem
      tags: [inventory]
      summary: Record a borrowed inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BorrowedItemInput"
      responses:
        "201":
          description: Created borrowed item record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BorrowedItem"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Not enough inventory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inventory/borrowed-items/{id}:
    patch:
      operationId: updateBorrowedItem
      tags: [inventory]
      summary: Update a borrowed item record
      parameters:
        - $ref: "#/components/parameters/BorrowedItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BorrowedItemInput"
      responses:
        "200":
          description: Updated borrowed item record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BorrowedItem"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Not enough inventory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBorrowedItem
      tags: [inventory]
      summary: Delete a borrowed item record
      parameters:
        - $ref: "#/components/parameters/BorrowedItemId"
      responses:
        "204":
          description: Borrowed item record deleted
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    InventoryItemId:
      name: id
      in: path
      required: true
      schema:
        type: string
    BorrowedItemId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    InventoryItem:
      type: object
      required: [id, name, category, quantity, unit, threshold, location]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string, enum: [Paper, Cards, Finishing, Vinyl, Ink, Office, Tools, Packaging, Safety] }
        quantity: { type: integer, minimum: 0 }
        unit: { type: string }
        threshold: { type: integer, minimum: 0 }
        location: { type: string }
        note: { type: ["string", "null"] }
    InventoryItemInput:
      type: object
      required: [id, name, category, quantity, unit, threshold, location]
      properties:
        id: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        category: { type: string, enum: [Paper, Cards, Finishing, Vinyl, Ink, Office, Tools, Packaging, Safety] }
        quantity: { type: integer, minimum: 0 }
        unit: { type: string, minLength: 1 }
        threshold: { type: integer, minimum: 0 }
        location: { type: string, minLength: 1 }
        note: { type: ["string", "null"] }
    StockActivity:
      type: object
      required: [id, itemId, itemName, action, amount, quantityAfter, unit, timestamp]
      properties:
        id: { type: string }
        itemId: { type: string }
        itemName: { type: string }
        action: { type: string, enum: [take, stock] }
        amount: { type: integer, minimum: 1 }
        quantityAfter: { type: integer, minimum: 0 }
        unit: { type: string }
        timestamp: { type: string, format: date-time }
    AuditRecord:
      type: object
      required: [id, date, itemId, itemName, quantity, unit, requesterName, department, purpose, createdAt]
      properties:
        id: { type: string }
        date: { type: string, format: date }
        itemId: { type: string }
        itemName: { type: string }
        quantity: { type: integer, minimum: 1 }
        unit: { type: string }
        requesterName: { type: string }
        department: { type: string }
        purpose: { type: string }
        createdAt: { type: string, format: date-time }
    AuditRecordInput:
      type: object
      required: [date, itemId, quantity, requesterName, department, purpose]
      properties:
        date: { type: string, format: date }
        itemId: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 1 }
        requesterName: { type: string, minLength: 1 }
        department: { type: string, minLength: 1 }
        purpose: { type: string, minLength: 1 }
    InventoryAdjustmentInput:
      type: object
      required: [itemId, quantity]
      properties:
        itemId: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 0 }
    InventoryBootstrapInput:
      type: object
      required: [items, activities, auditRecords, borrowedItems]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/InventoryItemInput" }
        activities:
          type: array
          items: { $ref: "#/components/schemas/StockActivity" }
        auditRecords:
          type: array
          items: { $ref: "#/components/schemas/AuditRecordInput" }
        borrowedItems:
          type: array
          items: { $ref: "#/components/schemas/BorrowedItemInput" }
    InventoryImportInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/InventoryItemInput" }
    InventoryImportResponse:
      type: object
      required: [insertedCount, items]
      properties:
        insertedCount: { type: integer, minimum: 0 }
        items:
          type: array
          items: { $ref: "#/components/schemas/InventoryItem" }
    BorrowedItem:
      type: object
      required: [id, dateBorrowed, borrowerName, itemId, itemName, quantity, unit, conditionBorrowed, dateReturned, conditionReturned, status, createdAt]
      properties:
        id: { type: string }
        dateBorrowed: { type: string, format: date }
        borrowerName: { type: string }
        itemId: { type: string }
        itemName: { type: string }
        quantity: { type: integer, minimum: 1 }
        unit: { type: string }
        conditionBorrowed: { type: string }
        dateReturned: { type: ["string", "null"], format: date }
        conditionReturned: { type: ["string", "null"] }
        status: { type: string, enum: [Borrowed, Broke, Returned] }
        createdAt: { type: string, format: date-time }
    BorrowedItemInput:
      type: object
      required: [dateBorrowed, borrowerName, itemId, quantity, conditionBorrowed, status]
      properties:
        dateBorrowed: { type: string, format: date }
        borrowerName: { type: string, minLength: 1 }
        itemId: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 1 }
        conditionBorrowed: { type: string, minLength: 1 }
        dateReturned: { type: ["string", "null"], format: date }
        conditionReturned: { type: ["string", "null"] }
        status: { type: string, enum: [Borrowed, Broke, Returned] }
    InventoryState:
      type: object
      required: [items, activities, auditRecords, borrowedItems]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/InventoryItem" }
        activities:
          type: array
          items: { $ref: "#/components/schemas/StockActivity" }
        auditRecords:
          type: array
          items: { $ref: "#/components/schemas/AuditRecord" }
        borrowedItems:
          type: array
          items: { $ref: "#/components/schemas/BorrowedItem" }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"