# lib/api-zod/src/generated/types/inventoryImportInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryItemInput } from './inventoryItemInput';

export interface InventoryImportInput {
  /** @minItems 1 */
  items: InventoryItemInput[];
}